move stata helper functions to their own file
