+++ pystata-kernel/utils.py
@@ -67,3 +67,5 @@ def mac_find_path():
         return ''
     else:
         return str(path)
+
+
